Moved the Player and Team classes out of the Players class
