+++ tichu-model/src/main/java/net/incongru/tichu/model/Players.java
@@ -5,13 +5,10 @@ import com.google.common.collect.Iterators;
 import com.google.common.collect.PeekingIterator;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.Iterator;
-import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
 import java.util.stream.Stream;
 
 import static java.util.function.Function.identity;
@@ -84,96 +81,4 @@ public class Players {
         return teams[i - 1];
     }
 
-    static public class Team {
-        private final String name;
-        private final Player player1;
-        private final Player player2;
-
-        public Team(String name, Player player1, Player player2) {
-            Preconditions.checkNotNull(name, "Team name can't be null");
-            Preconditions.checkNotNull(player1, "Player 1 for team %s can't be null", name);
-            Preconditions.checkNotNull(player2, "Player 2 for team %s can't be null", name);
-            this.name = name;
-            this.player1 = player1;
-            this.player2 = player2;
-        }
-
-        public String getName() {
-            return name;
-        }
-
-        @Override
-        public String toString() {
-            return "Team{" +
-                    "name='" + name + '\'' +
-                    ", player1=" + player1 +
-                    ", player2=" + player2 +
-                    '}';
-        }
-    }
-
-    static public class Player {
-        private final String name;
-        private final Set<Card> hand = new LinkedHashSet<>();
-        private final Set<Card> wonCards = new LinkedHashSet<>();
-
-        public Player(String name) {
-            Preconditions.checkNotNull(name, "Player name can't be null");
-            this.name = name;
-        }
-
-        public String name() {
-            return name;
-        }
-
-        public void reclaimCards() {
-            hand.clear();
-            wonCards.clear();
-        }
-
-        public Set<Card> hand() {
-            return Collections.unmodifiableSet(hand);
-        }
-
-        /**
-         * The given card is dealt to this player.
-         * OhMyTenses!
-         */
-        public void deal(Card card) {
-            hand.add(card);
-        }
-
-        /**
-         * This player picks up the given cards from the trick.
-         * // TODO we may want to keep track of the various tricks instead of flattening ?
-         */
-        public void pickup(Set<Card> cards) {
-            wonCards.addAll(cards);
-        }
-
-        public Set<Card> wonCards() {
-            return Collections.unmodifiableSet(wonCards);
-        }
-
-        /**
-         * The player has succesfully played these, we remove them from her hand.
-         */
-        void discard(Set<Card> cards) {
-            if (!hand.containsAll(cards)) {
-                throw new IllegalStateException("Could not remove cards " + cards + " from player hand " + hand);
-            }
-            // The boolean returned by removeAll does not indicate success, but rather that the collection was mutated
-            // So it's false if cards is empty, OR if cards contains at least one card in hand, but doesn't validate all were in hand
-            hand.removeAll(cards);
-        }
-
-        @Override
-        public String toString() {
-            return "Player{" +
-                    "name='" + name + '\'' +
-                    ", hand=" + hand +
-                    ", wonCards=" + wonCards +
-                    '}';
-        }
-    }
 }
